ENH: Allow to define various cased variants via :SwapList...Case commands and config
The :SwapListLowerCase, :SwapListUpperCase, :SwapListTitleCase, :SwapListSmartCase variants of :SwapList add the corresponding variant.
Likewise, a {'case': '...'} configuration offers the same for defaults.
This also simplifies the g:default_swap_list considerably.
For the implementation, split off s:GetListsFromScope() and make s:GetLists() check the new 'case' configuration, and handle the transformations generically via Funcrefs that are passed to a new s:Transform() function, which converts the 'name' and 'options' of a copy of the swap list. This also checks for actual modifications (in case of single letter swaps, title-case and upper-case are identical), and suppresses duplicates.